--- public/src/components/channelManagement/bannerTests/bannerTestEditor.tsx
@@ -1,302 +1,0 @@
-import React from 'react';
-import { Region } from '../../../utils/models';
-import { ArticlesViewedSettings, DeviceType, setStatus, UserCohort } from '../helpers/shared';
-import { ARTICLE_COUNT_TEMPLATE } from '../helpers/validation';
-import { Typography } from '@material-ui/core';
-import BannerTestVariantEditor from './bannerTestVariantEditor';
-import TestVariantsEditor from '../testVariantsEditor';
-import TestEditorHeader from '../testEditorHeader';
-import TestEditorMinArticlesViewedInput from '../testEditorMinArticlesViewedInput';
-import TestEditorTargetAudienceSelector from '../testEditorTargetAudienceSelector';
-import TestEditorArticleCountEditor, {
-  DEFAULT_ARTICLES_VIEWED_SETTINGS,
-} from '../testEditorArticleCountEditor';
-import TestEditorActionButtons from '../testEditorActionButtons';
-import LiveSwitch from '../../shared/liveSwitch';
-import useValidation from '../hooks/useValidation';
-import { BannerContent, BannerTest, BannerVariant } from '../../../models/banner';
-import { getDefaultVariant } from './utils/defaults';
-import TestEditorVariantSummary from '../testEditorVariantSummary';
-import BannerVariantPreview from './bannerVariantPreview';
-import { ControlProportionSettings } from '../helpers/controlProportionSettings';
-import TestVariantsSplitEditor from '../testVariantsSplitEditor';
-import { useStyles } from '../helpers/testEditorStyles';
-
-const copyHasTemplate = (content: BannerContent, template: string): boolean =>
-  (content.heading && content.heading.includes(template)) ||
-  (content.paragraphs && content.paragraphs.some(para => para.includes(template))) ||
-  (content.messageText != null && content.messageText.includes(template));
-
-const testCopyHasTemplate = (test: BannerTest, template: string): boolean =>
-  test.variants.some(
-    variant =>
-      (variant.bannerContent && copyHasTemplate(variant.bannerContent, template)) ||
-      (variant.mobileBannerContent && copyHasTemplate(variant.mobileBannerContent, template)),
-  );
-
-interface BannerTestEditorProps {
-  test: BannerTest;
-  onChange: (updatedTest: BannerTest) => void;
-  onValidationChange: (isValid: boolean) => void;
-  visible: boolean;
-  editMode: boolean;
-  onDelete: () => void;
-  onArchive: () => void;
-  onTestSelected: (testName: string) => void;
-  testNames: string[];
-  testNicknames: string[];
-  createTest: (newTest: BannerTest) => void;
-}
-
-const BannerTestEditor: React.FC<BannerTestEditorProps> = ({
-  test,
-  onChange,
-  onValidationChange,
-  visible,
-  editMode,
-  onArchive,
-  onDelete,
-  onTestSelected,
-  testNames,
-  testNicknames,
-  createTest,
-}: BannerTestEditorProps) => {
-  const classes = useStyles();
-
-  const setValidationStatusForField = useValidation(onValidationChange);
-
-  const getArticlesViewedSettings = (test: BannerTest): ArticlesViewedSettings | undefined => {
-    if (!!test.articlesViewedSettings) {
-      return test.articlesViewedSettings;
-    }
-    if (testCopyHasTemplate(test, ARTICLE_COUNT_TEMPLATE)) {
-      return DEFAULT_ARTICLES_VIEWED_SETTINGS;
-    }
-    return undefined;
-  };
-
-  const updateTest = (updatedTest: BannerTest): void => {
-    onChange({
-      ...updatedTest,
-      // To save dotcom from having to work this out
-      articlesViewedSettings: getArticlesViewedSettings(updatedTest),
-    });
-  };
-
-  const onMinArticlesViewedValidationChanged = (isValid: boolean): void =>
-    setValidationStatusForField('minArticlesViewed', isValid);
-
-  const onArticlesViewedSettingsValidationChanged = (isValid: boolean): void =>
-    setValidationStatusForField('articlesViewedSettings', isValid);
-
-  const onVariantsSplitSettingsValidationChanged = (isValid: boolean): void =>
-    setValidationStatusForField('variantsSplitSettings', isValid);
-
-  const onControlProportionSettingsChange = (
-    controlProportionSettings?: ControlProportionSettings,
-  ): void => updateTest({ ...test, controlProportionSettings });
-
-  const onLiveSwitchChange = (isOn: boolean): void => {
-    updateTest({ ...test, isOn, status: setStatus(isOn) });
-  };
-
-  const onVariantsChange = (updatedVariantList: BannerVariant[]): void => {
-    updateTest({ ...test, variants: updatedVariantList });
-  };
-
-  const onVariantChange = (updatedVariant: BannerVariant): void => {
-    onVariantsChange(
-      test.variants.map(variant =>
-        variant.name === updatedVariant.name ? updatedVariant : variant,
-      ),
-    );
-  };
-
-  const onVariantDelete = (deletedVariantName: string): void => {
-    onVariantsChange(test.variants.filter(variant => variant.name !== deletedVariantName));
-  };
-
-  const onMinArticlesViewedChange = (updatedMinArticles: number): void => {
-    updateTest({
-      ...test,
-      minArticlesBeforeShowingBanner: updatedMinArticles,
-    });
-  };
-
-  const onRegionsChange = (updatedRegions: Region[]): void => {
-    updateTest({ ...test, locations: updatedRegions });
-  };
-
-  const onCohortChange = (updatedCohort: UserCohort): void => {
-    updateTest({ ...test, userCohort: updatedCohort });
-  };
-
-  const onDeviceTypeChange = (updatedDeviceType: DeviceType): void => {
-    updateTest({ ...test, deviceType: updatedDeviceType });
-  };
-
-  const onArticlesViewedSettingsChange = (
-    updatedArticlesViewedSettings?: ArticlesViewedSettings,
-  ): void => {
-    updateTest({
-      ...test,
-      articlesViewedSettings: updatedArticlesViewedSettings,
-    });
-  };
-
-  const onCopy = (name: string, nickname: string): void => {
-    onTestSelected(name);
-    createTest({ ...test, name: name, nickname: nickname, isOn: false, status: 'Draft' });
-  };
-
-  const renderVariantEditor = (variant: BannerVariant): React.ReactElement => (
-    <BannerTestVariantEditor
-      key={`banner-${test.name}-${variant.name}`}
-      variant={variant}
-      onVariantChange={onVariantChange}
-      onDelete={(): void => onVariantDelete(variant.name)}
-      editMode={editMode}
-      onValidationChange={(isValid: boolean): void =>
-        setValidationStatusForField(variant.name, isValid)
-      }
-    />
-  );
-
-  const renderVariantSummary = (variant: BannerVariant): React.ReactElement => (
-    <TestEditorVariantSummary
-      name={variant.name}
-      testName={test.name}
-      testType="BANNER"
-      isInEditMode={editMode}
-      topButton={<BannerVariantPreview variant={variant} />}
-      platform="DOTCOM" // hardcoded as banners are currently not supported in AMP or Apple News
-    />
-  );
-
-  const createVariant = (name: string): void => {
-    const newVariant: BannerVariant = {
-      ...getDefaultVariant(),
-      name: name,
-    };
-    onVariantsChange([...test.variants, newVariant]);
-  };
-
-  const onVariantClone = (originalVariant: BannerVariant, clonedVariantName: string): void => {
-    const newVariant: BannerVariant = {
-      ...originalVariant,
-      name: clonedVariantName,
-    };
-    onVariantsChange([...test.variants, newVariant]);
-  };
-
-  if (test && visible) {
-    return (
-      <div className={classes.container}>
-        <div className={classes.headerAndSwitchContainer}>
-          <TestEditorHeader name={test.name} nickname={test.nickname} />
-
-          <LiveSwitch
-            label="Live on Guardian.com"
-            isLive={test.isOn}
-            isDisabled={!editMode}
-            onChange={onLiveSwitchChange}
-          />
-        </div>
-
-        <div className={classes.sectionContainer}>
-          <Typography variant={'h3'} className={classes.sectionHeader}>
-            Variants
-          </Typography>
-          <div>
-            <TestVariantsEditor
-              variants={test.variants}
-              createVariant={createVariant}
-              testName={test.name}
-              editMode={editMode}
-              renderVariantEditor={renderVariantEditor}
-              renderVariantSummary={renderVariantSummary}
-              onVariantDelete={onVariantDelete}
-              onVariantClone={onVariantClone}
-            />
-          </div>
-        </div>
-
-        {test.variants.length > 1 && (
-          <div className={classes.sectionContainer}>
-            <Typography variant={'h3'} className={classes.sectionHeader}>
-              Variants split
-            </Typography>
-            <div>
-              <TestVariantsSplitEditor
-                variants={test.variants}
-                controlProportionSettings={test.controlProportionSettings}
-                onControlProportionSettingsChange={onControlProportionSettingsChange}
-                onValidationChange={onVariantsSplitSettingsValidationChanged}
-                isDisabled={!editMode}
-              />
-            </div>
-          </div>
-        )}
-
-        <div className={classes.sectionContainer}>
-          <Typography variant={'h3'} className={classes.sectionHeader}>
-            Display rules
-          </Typography>
-
-          <TestEditorMinArticlesViewedInput
-            minArticles={test.minArticlesBeforeShowingBanner}
-            isDisabled={!editMode}
-            onValidationChange={onMinArticlesViewedValidationChanged}
-            onUpdate={onMinArticlesViewedChange}
-          />
-        </div>
-
-        <div className={classes.sectionContainer}>
-          <Typography variant={'h3'} className={classes.sectionHeader}>
-            Target audience
-          </Typography>
-
-          <TestEditorTargetAudienceSelector
-            selectedRegions={test.locations}
-            onRegionsUpdate={onRegionsChange}
-            selectedCohort={test.userCohort}
-            onCohortChange={onCohortChange}
-            selectedDeviceType={test.deviceType ?? 'All'}
-            onDeviceTypeChange={onDeviceTypeChange}
-            isDisabled={!editMode}
-            showSupporterStatusSelector={true}
-            showDeviceTypeSelector={true}
-          />
-        </div>
-
-        <div className={classes.sectionContainer}>
-          <Typography variant={'h3'} className={classes.sectionHeader}>
-            Article count
-          </Typography>
-
-          <TestEditorArticleCountEditor
-            articlesViewedSettings={test.articlesViewedSettings}
-            onArticlesViewedSettingsChanged={onArticlesViewedSettingsChange}
-            onValidationChange={onArticlesViewedSettingsValidationChanged}
-            isDisabled={!editMode}
-          />
-        </div>
-        <div className={classes.buttonsContainer}>
-          <TestEditorActionButtons
-            existingNames={testNames}
-            sourceName={test.name}
-            existingNicknames={testNicknames}
-            sourceNickname={test.nickname}
-            isDisabled={!editMode}
-            onArchive={onArchive}
-            onDelete={onDelete}
-            onCopy={onCopy}
-          />
-        </div>
-      </div>
-    );
-  }
-  return null;
-};
-
-export default BannerTestEditor;
